+++ packages/cache/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @suspensive/cache
 
+## 0.3.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @suspensive/utils@2.12.3
+
 ## 0.3.5
 
 ### Patch Changes