+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @suspensive/react-native
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @suspensive/utils@2.12.3
+
 ## 0.0.4
 
 ### Patch Changes